edit my answer

+++ src/k_backjoon/A27433.java
@@ -5,11 +5,12 @@ import java.io.IOException;
 import java.io.InputStreamReader;
 
 class A27433 {
+	//맞았습니다!!
 	public static void main(String[] args) throws IOException {
 		BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
-		double N = Double.parseDouble(br.readLine());
-		double answer = 1;
-		for(double i = 2; i <= N; i++) answer *= i;
+		int N = Integer.parseInt(br.readLine());
+		long answer = 1;
+		for(int i = 2; i <= N; i++) answer *= i;
 		System.out.print(answer);
 	}
 }
